# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Mandara API
  version: 0.1.0
  description: |
    Mandara is a self-hosted light-novel reading server.

    This document is the single source of truth for the HTTP API contract:
    the backend implements it and the frontend generates its types from it
    (see `frontend/package.json`, script `api-types`).

    Model overview: `books` are pure metadata entries; `book_files` are the
    actual books (uploads or virtual plugin books). Several files can share
    one metadata entry. Chapters, reading sessions and shares attach to
    files. Every file is either `private` (owner + admins only) or `public`
    (all logged-in users).
servers:
  - url: http://127.0.0.1:8080
    description: Local development server (default)
...
paths:
...
  /api/series:
    get:
      tags: [series]
      summary: |
        List series visible to the caller: those created by the caller,
        and series owning at least one visible member file (admins see
        all). Each entry carries its member count.
      operationId: listSeries
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Visible series
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SeriesBrief"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [series]
      summary: Create a series (any logged-in user; the creator manages it)
      operationId: createSeries
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:
                  type: string
                authors:
                  type: array
                  items:
                    type: string
                description:
                  type: string
                ext:
                  $ref: "#/components/schemas/SeriesExt"
      responses:
        "201":
          description: Created series
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeriesBrief"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/series/{id}:
    get:
      tags: [series]
      summary: |
        Series detail: metadata plus member books in volume order (files
        filtered by visibility; entries without visible files still
        appear with an empty `files` list).
      operationId: getSeriesDetail
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SeriesId"
      responses:
        "200":
          description: Series detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeriesDetail"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [series]
      summary: Edit series metadata (creator or admin)
      operationId: patchSeries
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SeriesId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                authors:
                  type: array
                  items:
                    type: string
                description:
                  type: string
                ext:
                  description: Extended-metadata merge patch (absent = keep, null = clear, value = set)
                  $ref: "#/components/schemas/SeriesExt"
      responses:
        "200":
          description: Updated series
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeriesBrief"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [series]
      summary: |
        Delete a series (creator or admin). Member books are unassigned
        (their entries stay standalone); the series row is removed.
      operationId: deleteSeries
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SeriesId"
      responses:
        "204":
          description: Deleted
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/files/{id}/shares:
    post:
      tags: [shares]
      summary: Create a share link for a file or one of its sessions
      operationId: createShare
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/FileId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateShare"
      responses:
        "201":
          description: Share link created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShareResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
...
    FileId:
      name: id
      in: path
      required: true
      description: Book file id
      schema:
        type: string
    SeriesId:
      name: id
      in: path
      required: true
      description: Series id
      schema:
        type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Not allowed to perform this action
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource does not exist (or share expired)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Visibility:
      type: string
      enum: [private, public]
    BookMeta:
      type: object
      required: [id, title, authors, description, cover_url, ext, created_by, created_at, series_id, volume_no]
      properties:
        id:
          type: string
        title:
          type: string
        authors:
          type: array
          items:
            type: string
        description:
          type: [string, "null"]
        cover_url:
          type: [string, "null"]
        ext:
          $ref: "#/components/schemas/BookExt"
        created_by:
          description: Id of the user who created this metadata entry
          type: [string, "null"]
        created_at:
          type: string
          format: date-time
        series_id:
          description: |
            Series this book is a volume of (null = standalone). Volume
            splits assign every 卷 its own book entry plus one series
            row; manual assignment (`PATCH /api/books/{id}`) moves books
            in and out of series.
          type: [string, "null"]
        volume_no:
          description: 1-based volume ordinal inside the series (0 = standalone)
          type: integer
          minimum: 0
    BookExt:
      type: object
      description: |
        Extended book metadata (bangumi/douban-style). Every known key is
        optional; `null`/absent = unset. Unknown keys are preserved
        verbatim so future server fields round-trip through old clients.
        `isbn` is stored canonically as a hyphenless ISBN-13 (valid
        ISBN-10 inputs are converted); invalid ISBNs are rejected with
        400.
      properties:
        subtitle:
          description: 副标题
          type: [string, "null"]
        original_title:
          description: 原作名
          type: [string, "null"]
        isbn:
          description: Canonical hyphenless ISBN-13
          type: [string, "null"]
        publisher:
          description: 出版社
          type: [string, "null"]
        pub_date:
          description: 出版时间 (stored exactly as supplied)
          type: [string, "null"]
        translators:
          description: 译者
          type: [array, "null"]
          items:
            type: string
        illustrators:
          description: 插画师
          type: [array, "null"]
          items:
            type: string
        pages:
          description: 页数
          type: [integer, "null"]
          minimum: 0
        price:
          description: 定价 (currency as printed)
          type: [string, "null"]
        binding:
          description: 装帧 (文库/单行本/…)
          type: [string, "null"]
        language:
          description: BCP-47 language hint ("zh-CN", "ja")
          type: [string, "null"]
      additionalProperties: true
    SeriesExt:
      type: object
      description: |
        Extended series metadata: edition-independent facts of the
        publication family. Same rules as `BookExt` (all optional,
        unknown keys preserved).
      properties:
        original_title:
          description: 原作名
          type: [string, "null"]
        publisher:
          description: 出版社
          type: [string, "null"]
        pub_date:
          description: 首卷发售时间 (verbatim)
          type: [string, "null"]
        language:
          description: BCP-47 language hint
          type: [string, "null"]
        status:
          description: Publication status
          type: [string, "null"]
          enum: [ongoing, completed, hiatus]
        total_volumes:
          description: Planned volume count
          type: [integer, "null"]
          minimum: 0
        tags:
          description: 类型标签
          type: [array, "null"]
          items:
            type: string
      additionalProperties: true
    FileMeta:
      type: object
      required:
        - id
        - book_id
        - source
        - external_id
        - content_source
        - content_external_id
        - format
        - label
        - visibility
        - owner_id
        - chapter_count
        - created_at
        - volume_no
        - volume_offset
      properties:
        id:
          type: string
        book_id:
          type: string
        source:
          description: "'local' for uploads, otherwise the plugin id"
          type: string
        external_id:
          description: Id within the source (same as `id` for local files)
          type: string
        content_source:
          description: |
            Instance id whose chapters back this file when different from
            `source` (metadata/content source separation); null = `source`.
          type: string
          nullable: true
        content_external_id:
          description: |
            Book id inside the content source; null = `external_id`.
          type: string
          nullable: true
        format:
          type: string
          enum: [epub, txt, plugin]
        label:
          description: Short human label for this edition
          type: string
        visibility:
          $ref: "#/components/schemas/Visibility"
        owner_id:
          type: [string, "null"]
        chapter_count:
          type: integer
        created_at:
          type: string
          format: date-time
        volume_no:
          description: |
            Volume slice of the source book this file covers (0 = the
            whole book). A multi-volume plugin book is acquired as one
            file per 卷.
          type: integer
          minimum: 0
        volume_offset:
          description: |
            Flat chapter index of this volume's first chapter inside the
            source (persisted because the plugin is stateless); lazy
            chapter pulls use `volume_offset + idx`.
          type: integer
          minimum: 0
        original:
          description: |
            Retained original file bytes (`data/files/`, served by
            `GET /api/files/{id}/download`). null = plugin chapter-mode
            files and uploads made before original retention.
          type: object
          nullable: true
          required: [ext, size, sha256]
          properties:
            ext:
              description: File extension without the dot (epub / txt)
              type: string
            size:
              type: integer
            sha256:
              type: string
...
    SeriesBrief:
      type: object
      description: |
        A series: metadata grouping for the volumes of one publication
        family (e.g. a light-novel series split from a multi-volume
        plugin source). Member books carry `series_id` + `volume_no`.
      required: [id, title, authors, description, cover_url, ext, status, volume_count, created_by, created_at]
      properties:
        id:
          type: string
        title:
          type: string
        authors:
          type: array
          items:
            type: string
        description:
          type: [string, "null"]
        cover_url:
          type: [string, "null"]
        ext:
          $ref: "#/components/schemas/SeriesExt"
        status:
          description: Publication status projected from `ext.status` (null = unset)
          type: [string, "null"]
          enum: [ongoing, completed, hiatus]
        volume_count:
          description: Number of member books (all visibilities)
          type: integer
        created_by:
          type: [string, "null"]
        created_at:
          type: string
          format: date-time
    SeriesDetail:
      type: object
      required: [series, books]
      properties:
        series:
          $ref: "#/components/schemas/SeriesBrief"
        books:
          description: Member books in volume order, files filtered by visibility
          type: array
          items:
            $ref: "#/components/schemas/BookDetail"
...
    BookDetail:
      type: object
      required: [book, files, series]
      properties:
        book:
          $ref: "#/components/schemas/BookMeta"
        files:
          description: Files of this book visible to the caller (empty for a metadata-only entry)
          type: array
          items:
            $ref: "#/components/schemas/FileMeta"
        series:
          $ref: "#/components/schemas/SeriesBrief"
          nullable: true
...
    ShareKind:
      type: string
      enum: [book, session]
    CreateShare:
      type: object
      properties:
        kind:
          $ref: "#/components/schemas/ShareKind"
        session_id:
          type: string
          description: Required when kind is "session"
        expires_days:
          type: integer
          minimum: 1
    ShareResponse:
      type: object
      required:
        - token
        - url
        - kind
        - file_id
        - session_id
        - expires_at
        - created_at
      properties:
        token:
          type: string
        url:
          description: Frontend path of the share page
          type: string
        kind:
          $ref: "#/components/schemas/ShareKind"
        file_id:
          type: string
        session_id:
          type: [string, "null"]
        expires_at:
          type: [string, "null"]
        created_at:
          type: string
          format: date-time
...